rename package and make variables public
